requirements.txt
openai==0.27.2
weaviate_client==3.15.4
python-dotenv==1.0.0
argparse==1.4.0
